Rapatriement dans la branche 0.3 des revisions 815 a 833.
git-svn-id: svn+ssh://svn.tuxfamily.org/svnroot/qet/qet/branches/0.3@834 bfdf4180-ca20-0410-9c96-a3a8aa849046

+++ sources/conductor.h
@@ -123,7 +123,7 @@ class Conductor : public QObject, public QGraphicsPathItem {
 	static QBrush square_brush;
 	static bool pen_and_brush_initialized;
 	/// facteur de taille du carre de saisie du segment
-	int segments_squares_scale_;
+	qreal segments_squares_scale_;
 	
 	private:
 	void segmentsToPath();
